refactor: calendarClient
